Consolidate the portfolio split into a single module
The PR should read as adding exactly one algorithm in one file. Fold the
split router's three files into `fynd-core/src/algorithm/split.rs`:
SplitAlgorithm now implements the portfolio directly (absorbing the former
ExpSplitAlgorithm), and the bounded candidate discovery (former
split_discovery.rs) becomes a clearly-marked section of the same file.

Remove the research scaffolding that made it read as several algorithms: the
SplitStrategy enum, the ExpSplitAlgorithm wrapper, and the split_incr /
split_ff entry points (recoverable from history). The portfolio's internal
floor / refined-disjoint / fill-and-spill allocation passes stay. Discovery
stays generic over the graph edge weight W (so tests exercise it on a bare
topology graph while production runs the DepthAndPrice graph) but drops the
SplitMarketRead abstraction for a concrete MarketDataView, its only caller.

Tests from the three files merge into one module (deduplicated). Delete the
split-comparison history page and its nav entry; history now lives in git log
and the PR description.
