+++ fynd-core/src/algorithm/mod.rs
@@ -22,8 +22,6 @@ pub mod bellman_ford;
 pub mod most_liquid;
 pub mod path_frank_wolfe;
 pub mod split;
-pub mod split_discovery;
-pub mod split_exp;
 #[allow(dead_code)]
 pub(crate) mod split_primitives;
 
@@ -38,7 +36,6 @@ pub use bellman_ford::BellmanFordAlgorithm;
 pub use most_liquid::MostLiquidAlgorithm;
 pub use path_frank_wolfe::PathFrankWolfeAlgorithm;
 pub use split::SplitAlgorithm;
-pub use split_exp::ExpSplitAlgorithm;
 use tycho_simulation::tycho_core::models::Address;
 
 use crate::{
